requirements.txt
httpx>=0.24.0
python-dotenv>=1.0.0
reflex==0.8.21
uvicorn[standard]==0.24.0
pydantic>=2.5.0
redis[asyncio]>=5.2.1,<8.0
asyncpg>=0.30.0
kafka-python==2.0.2


# Removed fastapi due to starlette version conflict with reflex